# openapi/atlas-api.yaml
openapi: 3.1.0
info:
  title: ATLAS API
  description: |
    API for ATLAS (Automated Tracking & Licensing Application System)
    Manages permit applications, documents, users, and audit logs.
  version: 1.0.0
  contact:
    name: ATLAS Development Team
    email: dev@atlas.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://localhost:5001
    description: Development server
  - url: https://api.atlas.example.com
    description: Production server
security:
  - EntraID: ['User.Read']
...
paths:
  /api/applications:
    get:
      tags: [Applications]
      summary: Get all applications
      description: Retrieve a list of applications with optional filtering
      operationId: getApplications
      parameters:
        - name: citizenId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by citizen ID
        - name: officerId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by assigned officer ID
        - name: status
          in: query
          schema:
            type: string
          description: Filter by application status
        - name: permitTypeId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by permit type ID
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by submission date (from)
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by submission date (to)
        - name: search
          in: query
          schema:
            type: string
          description: Search term for application number or citizen name
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationSummaryDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Citizen', 'Officer', 'Admin']
    
    post:
      tags: [Applications]
      summary: Submit new application
      description: Submit a new permit application
      operationId: submitApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitApplicationRequest'
            examples:
              standardApplication:
                summary: Standard permit application
                value:
                  citizenId: '12345678-1234-1234-1234-123456789012'
                  permitTypeId: '87654321-4321-4321-4321-210987654321'
                  citizenNotes: 'Need permit for business operations'
      responses:
        '201':
          description: Application created successfully
          content:
            application/json:
              schema:
                type: string
                format: uuid
              example: '11111111-2222-3333-4444-555555555555'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Citizen', 'Officer', 'Admin']
  /api/applications/{id}:
    get:
      tags: [Applications]
      summary: Get application by ID
      description: Retrieve detailed information about a specific application
      operationId: getApplicationById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Application ID
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetailDto'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Citizen', 'Officer', 'Admin']
  /api/applications/{id}/approve:
    post:
      tags: [Applications]
      summary: Approve application
      description: Approve a pending permit application
      operationId: approveApplication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Application ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveApplicationRequest'
      responses:
        '200':
          description: Application approved successfully
          content:
            application/json:
              schema:
                type: boolean
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Officer', 'Admin']
  /api/applications/{id}/reject:
    post:
      tags: [Applications]
      summary: Reject application
      description: Reject a pending permit application
      operationId: rejectApplication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Application ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectApplicationRequest'
      responses:
        '200':
          description: Application rejected successfully
          content:
            application/json:
              schema:
                type: boolean
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Officer', 'Admin']
  /api/applications/{id}/request-info:
    post:
      tags: [Applications]
      summary: Request additional information
      description: Request additional information from citizen for a pending application
      operationId: requestInfo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Application ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestInfoRequest'
      responses:
        '200':
          description: Information request sent successfully
          content:
            application/json:
              schema:
                type: boolean
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Officer', 'Admin']
  /api/applications/{id}/assign:
    post:
      tags: [Applications]
      summary: Assign application to officer
      description: Assign a pending application to a specific officer
      operationId: assignToOfficer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Application ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignToOfficerRequest'
      responses:
        '200':
          description: Application assigned successfully
          content:
            application/json:
              schema:
                type: boolean
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Officer', 'Admin']
...
  /api/auditlogs:
    get:
      tags: [AuditLogs]
      summary: Get audit logs
      description: Retrieve audit logs with optional filtering
      operationId: getAuditLogs
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by user ID
        - name: actionType
          in: query
          schema:
            type: string
          description: Filter by action type
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by date (from)
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by date (to)
        - name: recordId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by record ID
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLogDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Admin']
  /api/auditlogs/export:
    get:
      tags: [AuditLogs]
      summary: Export audit logs to CSV
      description: Export filtered audit logs to CSV format
      operationId: exportAuditLogs
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by user ID
        - name: actionType
          in: query
          schema:
            type: string
          description: Filter by action type
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by date (from)
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by date (to)
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
        - EntraID: ['Admin']
components:
  schemas:
    ApplicationSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '12345678-1234-1234-1234-123456789012'
        applicationNumber:
          type: string
          example: 'APP-2026-0001'
        status:
          type: integer
          format: int32
          description: Application status code
          example: 1
        submittedDate:
          type: string
          format: date-time
          nullable: true
        citizenId:
          type: string
          format: uuid
        permitTypeId:
          type: string
          format: uuid
        citizenName:
          type: string
          nullable: true
          example: 'John Doe'
        permitTypeName:
          type: string
          nullable: true
          example: 'Business Permit'
      required: [id, applicationNumber, status, citizenId, permitTypeId]
    ApplicationDetailDto:
      allOf:
        - $ref: '#/components/schemas/ApplicationSummaryDto'
        - type: object
          properties:
            reviewedDate:
              type: string
              format: date-time
              nullable: true
            citizenNotes:
              type: string
              default: ''
            officerNotes:
              type: string
              default: ''
            documents:
              type: array
              items:
                $ref: '#/components/schemas/DocumentDto'
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/ReviewDto'
            officerName:
              type: string
              nullable: true
              example: 'Jane Smith'
          required: [citizenNotes, officerNotes, documents, reviews]
    DocumentDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
          example: 'permit-application.pdf'
        contentType:
          type: string
          example: 'application/pdf'
        fileSize:
          type: integer
          format: int64
          example: 2048576
        blobUrl:
          type: string
          format: uri
          example: 'https://atlasstorage.blob.core.windows.net/documents/permit-application.pdf'
        uploadedDate:
          type: string
          format: date-time
      required: [id, fileName, contentType, fileSize, blobUrl, uploadedDate]
    ReviewDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        officerId:
          type: string
          format: uuid
        decision:
          type: integer
          format: int32
          description: 1=Approved, 2=Rejected
          enum: [1, 2]
        reasonCode:
          type: string
          nullable: true
          example: 'INCOMPLETE_DOCUMENTATION'
        comments:
          type: string
          nullable: true
        reviewedDate:
          type: string
          format: date-time
        isVisibleToCitizen:
          type: boolean
          default: false
      required: [id, officerId, decision, reviewedDate, isVisibleToCitizen]
...
    AuditLogDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
        actionType:
          type: string
          example: 'APPLICATION_SUBMITTED'
        timestamp:
          type: string
          format: date-time
        recordId:
          type: string
          format: uuid
        details:
          type: string
          nullable: true
      required: [id, actionType, timestamp, recordId]
    ProblemDetails:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
      required: [title, status]
    ValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
          required: [errors]
    SubmitApplicationRequest:
      type: object
      properties:
        citizenId:
          type: string
          format: uuid
        permitTypeId:
          type: string
          format: uuid
        citizenNotes:
          type: string
          default: ''
      required: [citizenId, permitTypeId]
    ApproveApplicationRequest:
      type: object
      properties:
        applicationId:
          type: string
          format: uuid
        officerId:
          type: string        
          format: uuid
        comments:
          type: string
          default: ''
      required: [applicationId]
    RejectApplicationRequest:
      type: object
      properties:
        applicationId:
          type: string
          format: uuid
        reasonCode:
          type: string
        comments:
          type: string
          nullable: true
      required: [applicationId, reasonCode]
    RequestInfoRequest:
      type: object
      properties:
        applicationId:
          type: string
          format: uuid
        officerId:
          type: string
          format: uuid
        message:
          type: string        
      required: [applicationId, message]
    AssignToOfficerRequest:
      type: object
      properties:
        applicationId:
          type: string
          format: uuid
        officerId:
          type: string
          format: uuid
      required: [applicationId, officerId]
...
  responses:
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetails'
  securitySchemes:
    EntraID:
      type: oauth2
      description: Microsoft Entra ID (Azure AD) authentication
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/token
          scopes:
            User.Read: Read user profile
            Citizen: Citizen role access
            Officer: Officer role access
            Admin: Admin role access